MTB01-GeLeagueInfo | Migrate code to typescript.
1. Add jest configuration for typescript.
2. Add typescript and tslint configs.
3. Ignore dist folder.
4. Migrate all javascript files from root to src folder.
5. Migrate all require statements to import statements.
6. Migrate all files from js to ts extension.
