--- lib/airtable.schema.js
@@ -1,16 +1,0 @@
-/* ––
- * –––– Schema declaration
- * –––––––––––––––––––––––––––––––– */
-const airtableSchema = {
-  Leagues: {
-    tableName: 'Leagues',
-    name: 'Name',
-    members: 'Members',
-    code: 'Code',
-  },
-};
-
-/* ––
- * –––– Exports
- * –––––––––––––––––––––––––––––––– */
-module.exports = airtableSchema;
