--- lib/tournament.js
@@ -1,11 +1,0 @@
-/* ––
- * –––– Public API
- * –––––––––––––––––––––––––––––––– */
-function getTournament(id) {
-  return `Tournament ${id}`;
-}
-
-/* ––
- * –––– Exports
- * –––––––––––––––––––––––––––––––– */
-module.exports = getTournament;
