+++ package.json
@@ -2,12 +2,13 @@
   "name": "masterboard-api",
   "version": "1.0.0-alpha.0",
   "description": "API implementation for Masterboard app.",
-  "main": "index.js",
+  "main": "dist/app.js",
   "author": "Gilberto López <glr2493@gmail.com>",
   "license": "MIT",
   "private": false,
   "scripts": {
-    "start": "node -r dotenv/config index.js dotenv_config_path=env/.env",
+    "build": "tsc",
+    "start": "node -r dotenv/config dist/app.js dotenv_config_path=env/.env",
     "test": "jest"
   },
   "dependencies": {
@@ -20,6 +21,12 @@
     "graphql-playground-middleware-express": "~1.7.12"
   },
   "devDependencies": {
-    "jest": "~25.1.0"
+    "@types/express": "~4.17.3",
+    "@types/jest": "~25.1.4",
+    "@types/node": "~13.9.2",
+    "jest": "~25.1.0",
+    "ts-jest": "~25.2.1",
+    "tslint": "~6.1.0",
+    "typescript": "~3.8.3"
   }
 }
