+++ src/lib/airtable.schema.ts
@@ -1,0 +1,9 @@
+/* ––
+ * –––– Schema declaration
+ * –––––––––––––––––––––––––––––––– */
+export const Leagues = {
+  tableName: 'Leagues',
+  name: 'Name',
+  members: 'Members',
+  code: 'Code',
+};
