+++ src/lib/airtable.service.ts
@@ -2,10 +2,10 @@
  * –––– Imports
  * –––––––––––––––––––––––––––––––– */
 // Third-party imports
-const Airtable = require('airtable');
+import Airtable from 'airtable';
 
 // App imports
-const { Leagues } = require('./airtable.schema');
+import { Leagues } from './airtable.schema';
 
 /* ––
  * –––– Datasource connection
@@ -15,21 +15,12 @@ const masterboardBase = Airtable.base(process.env.AIRTABLE_MASTERBOARD_BASE);
 /* ––
  * –––– API methods
  * –––––––––––––––––––––––––––––––– */
-const getLeagues = () => {
+export const getLeagues = () => {
   return masterboardBase(Leagues.tableName)
     .select()
     .all();
 };
 
-const getLeague = id => {
+export const getLeague = id => {
   return masterboardBase(Leagues.tableName).find(id);
-};
-
-/* ––
- * –––– Exports
- * –––––––––––––––––––––––––––––––– */
-
-module.exports = {
-  getLeagues,
-  getLeague,
 };
