+++ src/lib/tournament.ts
@@ -1,0 +1,6 @@
+/* ––
+ * –––– Public API
+ * –––––––––––––––––––––––––––––––– */
+export function getTournament(id: string) {
+  return `Tournament ${id}`;
+}
