+++ src/setup/environment-variables.ts
@@ -2,7 +2,7 @@
  * –––– Imports
  * –––––––––––––––––––––––––––––––– */
 // Third-party imports
-const colors = require('colors');
+import 'colors';
 
 /* ––
  * –––– Constants declaration
@@ -17,7 +17,7 @@ const environmentVariables = new Map([
 /* ––
  * –––– Utilities declaration
  * –––––––––––––––––––––––––––––––– */
-function checkEnvironmentVariables() {
+export function checkEnvironmentVariables() {
   console.log('\n⚙️ Checking environment variables\n'.underline.green.bold);
 
   const checkResults = Array.from(environmentVariables.entries()).map(
@@ -26,10 +26,10 @@ function checkEnvironmentVariables() {
       const headerPrefix = `  🧩  ${environmentVariable}:`;
 
       if (!value && required) {
-        console.log(`${headerPrefix} missing 🚨 (required)`.brightRed);
+        console.log(`${headerPrefix} missing 🚨 (required)`.red);
         return false;
       } else if (!value) {
-        console.log(`${headerPrefix} missing ⚠️ (optional)`.brightYellow);
+        console.log(`${headerPrefix} missing ⚠️ (optional)`.yellow);
       } else {
         console.log(`${headerPrefix} found ✔`.green);
       }
@@ -40,8 +40,3 @@ function checkEnvironmentVariables() {
   console.log('');
   return checkResults.every(value => value);
 }
-
-/* ––
- * –––– Exports
- * –––––––––––––––––––––––––––––––– */
-module.exports = checkEnvironmentVariables;
